requirements.txt
# requirements.txt
streamlit==1.26.0
qrcode==7.3.1
numpy==1.26.0
opencv-python-headless==4.8.1.78
Pillow==10.0.0
face-recognition==1.3.0
pyzbar==0.1.8
